requirements.txt
# ==================== requirements.txt ====================
# Core application
streamlit>=1.30.0
python-dotenv>=1.0.0
numpy>=1.26.0

# PDF processing
PyPDF2>=3.0.0

# Embeddings & retrieval
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# LLM provider (free-tier, fast inference)
groq>=0.9.0
